rf#260  Rename "newcontext" to "context"
