+++ catalog/context/__init__.py
@@ -1,0 +1,2 @@
+from . import category, pages, products, tags
+
