--- catalog/newcontext/__init__.py
@@ -1,3 +1,0 @@
-from . import category, pages, products, tags
-from .context import Context, ModelContext, Products, Tags
-
