+++ refarm_pagination/context.py
@@ -5,7 +5,7 @@ from refarm_pagination.pagination import Paginator, NeighborPages
 #  It's not clear now.
 #  For example client should know the class context details
 #  to fetch objects iterable.
-#  See `catalog.newcontext.products.PaginatedProducts#qs`.
+#  See `catalog.context.products.PaginatedProducts#qs`.
 class PaginationContext:
 
     def __init__(self, url, number, per_page, objects):
